Implement ExtractStructFile API
