+++ fs/fs.go
@@ -2,7 +2,6 @@ package fs
 
 import (
 	"fmt"
-	"os"
 	"sync"
 	"time"
 
@@ -128,120 +127,6 @@ func (fs *FileSystem) GetTransferMetrics() types.TransferMetrics {
 	return fs.session.GetTransferMetrics()
 }
 
-// ListGroupUsers lists all users in a group
-func (fs *FileSystem) ListGroupUsers(group string) ([]*types.IRODSUser, error) {
-	// check cache first
-	cachedUsers := fs.cache.GetGroupUsersCache(group)
-	if cachedUsers != nil {
-		return cachedUsers, nil
-	}
-
-	// otherwise, retrieve it and add it to cache
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	users, err := irods_fs.ListGroupUsers(conn, group)
-	if err != nil {
-		return nil, err
-	}
-
-	// cache it
-	fs.cache.AddGroupUsersCache(group, users)
-
-	return users, nil
-}
-
-// ListGroups lists all groups
-func (fs *FileSystem) ListGroups() ([]*types.IRODSUser, error) {
-	// check cache first
-	cachedGroups := fs.cache.GetGroupsCache()
-	if cachedGroups != nil {
-		return cachedGroups, nil
-	}
-
-	// otherwise, retrieve it and add it to cache
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	groups, err := irods_fs.ListGroups(conn)
-	if err != nil {
-		return nil, err
-	}
-
-	// cache it
-	fs.cache.AddGroupsCache(groups)
-
-	return groups, nil
-}
-
-// ListUserGroups lists all groups that a user belongs to
-func (fs *FileSystem) ListUserGroups(user string) ([]*types.IRODSUser, error) {
-	// check cache first
-	cachedGroups := fs.cache.GetUserGroupsCache(user)
-	if cachedGroups != nil {
-		return cachedGroups, nil
-	}
-
-	// otherwise, retrieve it and add it to cache
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	groupNames, err := irods_fs.ListUserGroupNames(conn, user)
-	if err != nil {
-		return nil, err
-	}
-
-	groups := []*types.IRODSUser{}
-	for _, groupName := range groupNames {
-		group, err := irods_fs.GetGroup(conn, groupName)
-		if err != nil {
-			return nil, err
-		}
-
-		groups = append(groups, group)
-	}
-
-	// cache it
-	fs.cache.AddUserGroupsCache(user, groups)
-
-	return groups, nil
-}
-
-// ListUsers lists all users
-func (fs *FileSystem) ListUsers() ([]*types.IRODSUser, error) {
-	// check cache first
-	cachedUsers := fs.cache.GetUsersCache()
-	if cachedUsers != nil {
-		return cachedUsers, nil
-	}
-
-	// otherwise, retrieve it and add it to cache
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	users, err := irods_fs.ListUsers(conn)
-	if err != nil {
-		return nil, err
-	}
-
-	// cache it
-	fs.cache.AddUsersCache(users)
-
-	return users, nil
-}
-
 // Stat returns file status
 func (fs *FileSystem) Stat(path string) (*Entry, error) {
 	irodsPath := util.GetCorrectIRODSPath(path)
@@ -344,215 +229,6 @@ func (fs *FileSystem) ExistsFile(path string) bool {
 	return entry.ID > 0
 }
 
-// ListACLs returns ACLs
-func (fs *FileSystem) ListACLs(path string) ([]*types.IRODSAccess, error) {
-	stat, err := fs.Stat(path)
-	if err != nil {
-		return nil, err
-	}
-
-	if stat.Type == DirectoryEntry {
-		return fs.ListDirACLs(path)
-	} else if stat.Type == FileEntry {
-		return fs.ListFileACLs(path)
-	}
-
-	return nil, fmt.Errorf("unknown type - %s", stat.Type)
-}
-
-// ListACLsForEntries returns ACLs for entries in a collection
-func (fs *FileSystem) ListACLsForEntries(path string) ([]*types.IRODSAccess, error) {
-	irodsPath := util.GetCorrectIRODSPath(path)
-
-	collectionEntry, err := fs.getCollection(irodsPath)
-	if err != nil {
-		return nil, err
-	}
-
-	collection := fs.getCollectionFromEntry(collectionEntry)
-
-	return fs.listACLsForEntries(collection)
-}
-
-// ListACLsWithGroupUsers returns ACLs
-func (fs *FileSystem) ListACLsWithGroupUsers(path string) ([]*types.IRODSAccess, error) {
-	stat, err := fs.Stat(path)
-	if err != nil {
-		return nil, err
-	}
-
-	accesses := []*types.IRODSAccess{}
-	if stat.Type == DirectoryEntry {
-		accessList, err := fs.ListDirACLsWithGroupUsers(path)
-		if err != nil {
-			return nil, err
-		}
-
-		accesses = append(accesses, accessList...)
-	} else if stat.Type == FileEntry {
-		accessList, err := fs.ListFileACLsWithGroupUsers(path)
-		if err != nil {
-			return nil, err
-		}
-
-		accesses = append(accesses, accessList...)
-	} else {
-		return nil, fmt.Errorf("unknown type - %s", stat.Type)
-	}
-
-	return accesses, nil
-}
-
-// ListDirACLs returns ACLs of a directory
-func (fs *FileSystem) ListDirACLs(path string) ([]*types.IRODSAccess, error) {
-	irodsPath := util.GetCorrectIRODSPath(path)
-
-	// check cache first
-	cachedAccesses := fs.cache.GetACLsCache(irodsPath)
-	if cachedAccesses != nil {
-		return cachedAccesses, nil
-	}
-
-	// otherwise, retrieve it and add it to cache
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	accesses, err := irods_fs.ListCollectionAccesses(conn, irodsPath)
-	if err != nil {
-		return nil, err
-	}
-
-	// cache it
-	fs.cache.AddACLsCache(irodsPath, accesses)
-
-	return accesses, nil
-}
-
-// ListDirACLsWithGroupUsers returns ACLs of a directory
-// CAUTION: this can fail if a group contains a lot of users
-func (fs *FileSystem) ListDirACLsWithGroupUsers(path string) ([]*types.IRODSAccess, error) {
-	accesses, err := fs.ListDirACLs(path)
-	if err != nil {
-		return nil, err
-	}
-
-	newAccesses := []*types.IRODSAccess{}
-	newAccessesMap := map[string]*types.IRODSAccess{}
-
-	for _, access := range accesses {
-		if access.UserType == types.IRODSUserRodsGroup {
-			// retrieve all users in the group
-			users, err := fs.ListGroupUsers(access.UserName)
-			if err != nil {
-				return nil, err
-			}
-
-			for _, user := range users {
-				userAccess := &types.IRODSAccess{
-					Path:        access.Path,
-					UserName:    user.Name,
-					UserZone:    user.Zone,
-					UserType:    user.Type,
-					AccessLevel: access.AccessLevel,
-				}
-
-				// remove duplicates
-				newAccessesMap[fmt.Sprintf("%s||%s", user.Name, access.AccessLevel)] = userAccess
-			}
-		} else {
-			newAccessesMap[fmt.Sprintf("%s||%s", access.UserName, access.AccessLevel)] = access
-		}
-	}
-
-	// convert map to array
-	for _, access := range newAccessesMap {
-		newAccesses = append(newAccesses, access)
-	}
-
-	return newAccesses, nil
-}
-
-// ListFileACLs returns ACLs of a file
-func (fs *FileSystem) ListFileACLs(path string) ([]*types.IRODSAccess, error) {
-	irodsPath := util.GetCorrectIRODSPath(path)
-
-	// check cache first
-	cachedAccesses := fs.cache.GetACLsCache(irodsPath)
-	if cachedAccesses != nil {
-		return cachedAccesses, nil
-	}
-
-	// otherwise, retrieve it and add it to cache
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	collectionEntry, err := fs.getCollection(util.GetIRODSPathDirname(irodsPath))
-	if err != nil {
-		return nil, err
-	}
-
-	collection := fs.getCollectionFromEntry(collectionEntry)
-
-	accesses, err := irods_fs.ListDataObjectAccesses(conn, collection, util.GetIRODSPathFileName(irodsPath))
-	if err != nil {
-		return nil, err
-	}
-
-	// cache it
-	fs.cache.AddACLsCache(irodsPath, accesses)
-
-	return accesses, nil
-}
-
-// ListFileACLsWithGroupUsers returns ACLs of a file
-func (fs *FileSystem) ListFileACLsWithGroupUsers(path string) ([]*types.IRODSAccess, error) {
-	accesses, err := fs.ListFileACLs(path)
-	if err != nil {
-		return nil, err
-	}
-
-	newAccesses := []*types.IRODSAccess{}
-	newAccessesMap := map[string]*types.IRODSAccess{}
-
-	for _, access := range accesses {
-		if access.UserType == types.IRODSUserRodsGroup {
-			// retrieve all users in the group
-			users, err := fs.ListGroupUsers(access.UserName)
-			if err != nil {
-				return nil, err
-			}
-
-			for _, user := range users {
-				userAccess := &types.IRODSAccess{
-					Path:        access.Path,
-					UserName:    user.Name,
-					UserZone:    user.Zone,
-					UserType:    user.Type,
-					AccessLevel: access.AccessLevel,
-				}
-
-				// remove duplicates
-				newAccessesMap[fmt.Sprintf("%s||%s", user.Name, access.AccessLevel)] = userAccess
-			}
-		} else {
-			newAccessesMap[fmt.Sprintf("%s||%s", access.UserName, access.AccessLevel)] = access
-		}
-	}
-
-	// convert map to array
-	for _, access := range newAccessesMap {
-		newAccesses = append(newAccesses, access)
-	}
-
-	return newAccesses, nil
-}
-
 // List lists all file system entries under the given path
 func (fs *FileSystem) List(path string) ([]*Entry, error) {
 	irodsPath := util.GetCorrectIRODSPath(path)
@@ -565,11 +241,6 @@ func (fs *FileSystem) List(path string) ([]*Entry, error) {
 	collection := fs.getCollectionFromEntry(collectionEntry)
 
 	return fs.listEntries(collection)
-}
-
-// SearchByMeta searches all file system entries with given metadata
-func (fs *FileSystem) SearchByMeta(metaname string, metavalue string) ([]*Entry, error) {
-	return fs.searchEntriesByMeta(metaname, metavalue)
 }
 
 // RemoveDir deletes a directory
@@ -950,286 +621,6 @@ func (fs *FileSystem) ReplicateFile(path string, resource string, update bool) e
 	return nil
 }
 
-// DownloadFile downloads a file to local
-func (fs *FileSystem) DownloadFile(irodsPath string, resource string, localPath string) error {
-	irodsSrcPath := util.GetCorrectIRODSPath(irodsPath)
-	localDestPath := util.GetCorrectIRODSPath(localPath)
-
-	localFilePath := localDestPath
-
-	srcStat, err := fs.Stat(irodsSrcPath)
-	if err != nil {
-		return types.NewFileNotFoundErrorf("could not find a data object")
-	}
-
-	if srcStat.Type == DirectoryEntry {
-		return fmt.Errorf("cannot download a collection %s", irodsSrcPath)
-	}
-
-	destStat, err := os.Stat(localDestPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			// file not exists, it's a file
-			// pass
-		} else {
-			return err
-		}
-	} else {
-		if destStat.IsDir() {
-			irodsFileName := util.GetIRODSPathFileName(irodsSrcPath)
-			localFilePath = util.MakeIRODSPath(localDestPath, irodsFileName)
-		} else {
-			return fmt.Errorf("file %s already exists", localDestPath)
-		}
-	}
-
-	return irods_fs.DownloadDataObject(fs.session, irodsSrcPath, resource, localFilePath)
-}
-
-// DownloadFileParallel downloads a file to local in parallel
-func (fs *FileSystem) DownloadFileParallel(irodsPath string, resource string, localPath string, taskNum int) error {
-	irodsSrcPath := util.GetCorrectIRODSPath(irodsPath)
-	localDestPath := util.GetCorrectIRODSPath(localPath)
-
-	localFilePath := localDestPath
-
-	srcStat, err := fs.Stat(irodsSrcPath)
-	if err != nil {
-		return types.NewFileNotFoundErrorf("could not find a data object")
-	}
-
-	if srcStat.Type == DirectoryEntry {
-		return fmt.Errorf("cannot download a collection %s", irodsSrcPath)
-	}
-
-	destStat, err := os.Stat(localDestPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			// file not exists, it's a file
-			// pass
-		} else {
-			return err
-		}
-	} else {
-		if destStat.IsDir() {
-			irodsFileName := util.GetIRODSPathFileName(irodsSrcPath)
-			localFilePath = util.MakeIRODSPath(localDestPath, irodsFileName)
-		} else {
-			return fmt.Errorf("file %s already exists", localDestPath)
-		}
-	}
-
-	return irods_fs.DownloadDataObjectParallel(fs.session, irodsSrcPath, resource, localFilePath, srcStat.Size, taskNum)
-}
-
-// DownloadFileParallelInBlocksAsync downloads a file to local in parallel
-func (fs *FileSystem) DownloadFileParallelInBlocksAsync(irodsPath string, resource string, localPath string, blockLength int64, taskNum int) (chan int64, chan error) {
-	irodsSrcPath := util.GetCorrectIRODSPath(irodsPath)
-	localDestPath := util.GetCorrectIRODSPath(localPath)
-
-	localFilePath := localDestPath
-
-	outputChan := make(chan int64, 1)
-	errChan := make(chan error, 1)
-
-	srcStat, err := fs.Stat(irodsSrcPath)
-	if err != nil {
-		errChan <- types.NewFileNotFoundErrorf("could not find a data object")
-		close(outputChan)
-		close(errChan)
-		return outputChan, errChan
-	}
-
-	if srcStat.Type == DirectoryEntry {
-		errChan <- fmt.Errorf("cannot download a collection %s", irodsSrcPath)
-		close(outputChan)
-		close(errChan)
-		return outputChan, errChan
-	}
-
-	destStat, err := os.Stat(localDestPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			// file not exists, it's a file
-			// pass
-		} else {
-			errChan <- err
-			close(outputChan)
-			close(errChan)
-			return outputChan, errChan
-		}
-	} else {
-		if destStat.IsDir() {
-			irodsFileName := util.GetIRODSPathFileName(irodsSrcPath)
-			localFilePath = util.MakeIRODSPath(localDestPath, irodsFileName)
-		} else {
-			errChan <- fmt.Errorf("file %s already exists", localDestPath)
-			close(outputChan)
-			close(errChan)
-			return outputChan, errChan
-		}
-	}
-
-	return irods_fs.DownloadDataObjectParallelInBlocksAsync(fs.session, irodsSrcPath, resource, localFilePath, srcStat.Size, blockLength, taskNum)
-}
-
-// UploadFile uploads a local file to irods
-func (fs *FileSystem) UploadFile(localPath string, irodsPath string, resource string, replicate bool) error {
-	localSrcPath := util.GetCorrectIRODSPath(localPath)
-	irodsDestPath := util.GetCorrectIRODSPath(irodsPath)
-
-	irodsFilePath := irodsDestPath
-
-	stat, err := os.Stat(localSrcPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			// file not exists
-			return types.NewFileNotFoundError("could not find the local file")
-		}
-		return err
-	}
-
-	if stat.IsDir() {
-		return types.NewFileNotFoundError("The local file is a directory")
-	}
-
-	entry, err := fs.Stat(irodsDestPath)
-	if err != nil {
-		if !types.IsFileNotFoundError(err) {
-			return err
-		}
-	} else {
-		switch entry.Type {
-		case FileEntry:
-			// do nothing
-		case DirectoryEntry:
-			localFileName := util.GetIRODSPathFileName(localSrcPath)
-			irodsFilePath = util.MakeIRODSPath(irodsDestPath, localFileName)
-		default:
-			return fmt.Errorf("unknown entry type %s", entry.Type)
-		}
-	}
-
-	err = irods_fs.UploadDataObject(fs.session, localSrcPath, irodsFilePath, resource, replicate)
-	if err != nil {
-		return err
-	}
-
-	fs.invalidateCacheForFileCreate(irodsFilePath)
-	fs.cachePropagation.PropagateFileCreate(irodsFilePath)
-	return nil
-}
-
-// UploadFileParallel uploads a local file to irods in parallel
-func (fs *FileSystem) UploadFileParallel(localPath string, irodsPath string, resource string, taskNum int, replicate bool) error {
-	localSrcPath := util.GetCorrectIRODSPath(localPath)
-	irodsDestPath := util.GetCorrectIRODSPath(irodsPath)
-
-	irodsFilePath := irodsDestPath
-
-	srcStat, err := os.Stat(localSrcPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			// file not exists
-			return types.NewFileNotFoundError("could not find the local file")
-		}
-		return err
-	}
-
-	if srcStat.IsDir() {
-		return types.NewFileNotFoundError("The local file is a directory")
-	}
-
-	destStat, err := fs.Stat(irodsDestPath)
-	if err != nil {
-		if !types.IsFileNotFoundError(err) {
-			return err
-		}
-	} else {
-		switch destStat.Type {
-		case FileEntry:
-			// do nothing
-		case DirectoryEntry:
-			localFileName := util.GetIRODSPathFileName(localSrcPath)
-			irodsFilePath = util.MakeIRODSPath(irodsDestPath, localFileName)
-		default:
-			return fmt.Errorf("unknown entry type %s", destStat.Type)
-		}
-	}
-
-	err = irods_fs.UploadDataObjectParallel(fs.session, localSrcPath, irodsFilePath, resource, taskNum, replicate)
-	if err != nil {
-		return err
-	}
-
-	fs.invalidateCacheForFileCreate(irodsFilePath)
-	fs.cachePropagation.PropagateFileCreate(irodsFilePath)
-	return nil
-}
-
-// UploadFileParallelInBlocksAsync uploads a local file to irods in parallel
-func (fs *FileSystem) UploadFileParallelInBlocksAsync(localPath string, irodsPath string, resource string, blockLength int64, taskNum int, replicate bool) (chan int64, chan error) {
-	localSrcPath := util.GetCorrectIRODSPath(localPath)
-	irodsDestPath := util.GetCorrectIRODSPath(irodsPath)
-
-	irodsFilePath := irodsDestPath
-
-	outputChan := make(chan int64, 1)
-	errChan := make(chan error, 1)
-
-	srcStat, err := os.Stat(localSrcPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			// file not exists
-			errChan <- types.NewFileNotFoundError("could not find the local file")
-			close(outputChan)
-			close(errChan)
-			return outputChan, errChan
-		}
-
-		errChan <- err
-		close(outputChan)
-		close(errChan)
-		return outputChan, errChan
-	}
-
-	if srcStat.IsDir() {
-		errChan <- types.NewFileNotFoundError("The local file is a directory")
-		close(outputChan)
-		close(errChan)
-		return outputChan, errChan
-	}
-
-	destStat, err := fs.Stat(irodsDestPath)
-	if err != nil {
-		if !types.IsFileNotFoundError(err) {
-			errChan <- err
-			close(outputChan)
-			close(errChan)
-			return outputChan, errChan
-		}
-	} else {
-		switch destStat.Type {
-		case FileEntry:
-			// do nothing
-		case DirectoryEntry:
-			localFileName := util.GetIRODSPathFileName(localSrcPath)
-			irodsFilePath = util.MakeIRODSPath(irodsDestPath, localFileName)
-		default:
-			errChan <- fmt.Errorf("unknown entry type %s", destStat.Type)
-			close(outputChan)
-			close(errChan)
-			return outputChan, errChan
-		}
-	}
-
-	outputChan2, errChan2 := irods_fs.UploadDataObjectParallelInBlockAsync(fs.session, localSrcPath, irodsFilePath, resource, blockLength, taskNum, replicate)
-
-	fs.invalidateCacheForFileCreate(irodsFilePath)
-	fs.cachePropagation.PropagateFileCreate(irodsFilePath)
-	return outputChan2, errChan2
-}
-
 // OpenFile opens an existing file for read/write
 func (fs *FileSystem) OpenFile(path string, resource string, mode string) (*FileHandle, error) {
 	irodsPath := util.GetCorrectIRODSPath(path)
@@ -1337,15 +728,6 @@ func (fs *FileSystem) CreateFile(path string, resource string, mode string) (*Fi
 	fs.cachePropagation.PropagateFileCreate(irodsPath)
 
 	return fileHandle, nil
-}
-
-// ClearCache clears all file system caches
-func (fs *FileSystem) ClearCache() {
-	fs.cache.ClearACLsCache()
-	fs.cache.ClearMetadataCache()
-	fs.cache.ClearEntryCache()
-	fs.cache.ClearNegativeEntryCache()
-	fs.cache.ClearDirCache()
 }
 
 // getCollectionNoCache returns collection entry
@@ -1510,106 +892,6 @@ func (fs *FileSystem) listEntries(collection *types.IRODSCollection) ([]*Entry, 
 	return entries, nil
 }
 
-// listACLsForEntries lists ACLs for entries in a collection
-func (fs *FileSystem) listACLsForEntries(collection *types.IRODSCollection) ([]*types.IRODSAccess, error) {
-	// check cache first
-	cachedAccesses := []*types.IRODSAccess{}
-	useCached := false
-
-	cachedDirEntryPaths := fs.cache.GetDirCache(collection.Path)
-	if cachedDirEntryPaths != nil {
-		useCached = true
-		for _, cachedDirEntryPath := range cachedDirEntryPaths {
-			cachedAccess := fs.cache.GetACLsCache(cachedDirEntryPath)
-			if cachedAccess != nil {
-				cachedAccesses = append(cachedAccesses, cachedAccess...)
-			} else {
-				useCached = false
-			}
-		}
-	}
-
-	if useCached {
-		return cachedAccesses, nil
-	}
-
-	// otherwise, retrieve it and add it to cache
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	collectionAccesses, err := irods_fs.ListAccessesForSubCollections(conn, collection.Path)
-	if err != nil {
-		return nil, err
-	}
-
-	accesses := []*types.IRODSAccess{}
-
-	accesses = append(accesses, collectionAccesses...)
-
-	// cache it
-	fs.cache.AddACLsCacheMulti(collectionAccesses)
-
-	dataobjectAccesses, err := irods_fs.ListAccessesForDataObjects(conn, collection)
-	if err != nil {
-		return nil, err
-	}
-
-	accesses = append(accesses, dataobjectAccesses...)
-
-	// cache it
-	fs.cache.AddACLsCacheMulti(dataobjectAccesses)
-
-	return accesses, nil
-}
-
-// searchEntriesByMeta searches entries by meta
-func (fs *FileSystem) searchEntriesByMeta(metaName string, metaValue string) ([]*Entry, error) {
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	collections, err := irods_fs.SearchCollectionsByMeta(conn, metaName, metaValue)
-	if err != nil {
-		return nil, err
-	}
-
-	entries := []*Entry{}
-
-	for _, coll := range collections {
-		entry := fs.getEntryFromCollection(coll)
-		entries = append(entries, entry)
-
-		// cache it
-		fs.cache.RemoveNegativeEntryCache(entry.Path)
-		fs.cache.AddEntryCache(entry)
-	}
-
-	dataobjects, err := irods_fs.SearchDataObjectsMasterReplicaByMeta(conn, metaName, metaValue)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, dataobject := range dataobjects {
-		if len(dataobject.Replicas) == 0 {
-			continue
-		}
-
-		entry := fs.getEntryFromDataObject(dataobject)
-		entries = append(entries, entry)
-
-		// cache it
-		fs.cache.RemoveNegativeEntryCache(entry.Path)
-		fs.cache.AddEntryCache(entry)
-	}
-
-	return entries, nil
-}
-
 // getDataObjectWithConnectionNoCache returns an entry for data object
 func (fs *FileSystem) getDataObjectWithConnectionNoCache(conn *connection.IRODSConnection, path string) (*Entry, error) {
 	// retrieve it and add it to cache
@@ -1700,320 +982,4 @@ func (fs *FileSystem) getDataObject(path string) (*Entry, error) {
 
 	// otherwise, retrieve it and add it to cache
 	return fs.getDataObjectNoCache(path)
-}
-
-// ListMetadata lists metadata for the given path
-func (fs *FileSystem) ListMetadata(path string) ([]*types.IRODSMeta, error) {
-	// check cache first
-	cachedEntry := fs.cache.GetMetadataCache(path)
-	if cachedEntry != nil {
-		return cachedEntry, nil
-	}
-
-	irodsCorrectPath := util.GetCorrectIRODSPath(path)
-
-	// otherwise, retrieve it and add it to cache
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	var metadataobjects []*types.IRODSMeta
-
-	if fs.ExistsDir(irodsCorrectPath) {
-		metadataobjects, err = irods_fs.ListCollectionMeta(conn, irodsCorrectPath)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		collectionEntry, err := fs.getCollection(util.GetIRODSPathDirname(path))
-		if err != nil {
-			return nil, err
-		}
-
-		collection := fs.getCollectionFromEntry(collectionEntry)
-
-		metadataobjects, err = irods_fs.ListDataObjectMeta(conn, collection, util.GetIRODSPathFileName(irodsCorrectPath))
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// cache it
-	fs.cache.AddMetadataCache(irodsCorrectPath, metadataobjects)
-
-	return metadataobjects, nil
-}
-
-// AddMetadata adds a metadata for the path
-func (fs *FileSystem) AddMetadata(irodsPath string, attName string, attValue string, attUnits string) error {
-	irodsCorrectPath := util.GetCorrectIRODSPath(irodsPath)
-
-	metadata := &types.IRODSMeta{
-		Name:  attName,
-		Value: attValue,
-		Units: attUnits,
-	}
-
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	if fs.ExistsDir(irodsCorrectPath) {
-		err = irods_fs.AddCollectionMeta(conn, irodsCorrectPath, metadata)
-		if err != nil {
-			return err
-		}
-	} else {
-		err = irods_fs.AddDataObjectMeta(conn, irodsCorrectPath, metadata)
-		if err != nil {
-			return err
-		}
-	}
-
-	fs.cache.RemoveMetadataCache(irodsCorrectPath)
-	return nil
-}
-
-// DeleteMetadata deletes a metadata for the path
-func (fs *FileSystem) DeleteMetadata(irodsPath string, attName string, attValue string, attUnits string) error {
-	irodsCorrectPath := util.GetCorrectIRODSPath(irodsPath)
-
-	metadata := &types.IRODSMeta{
-		Name:  attName,
-		Value: attValue,
-		Units: attUnits,
-	}
-
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	if fs.ExistsDir(irodsCorrectPath) {
-		err = irods_fs.DeleteCollectionMeta(conn, irodsCorrectPath, metadata)
-		if err != nil {
-			return err
-		}
-	} else {
-		err = irods_fs.DeleteDataObjectMeta(conn, irodsCorrectPath, metadata)
-		if err != nil {
-			return err
-		}
-	}
-
-	fs.cache.RemoveMetadataCache(irodsCorrectPath)
-	return nil
-}
-
-// invalidateCacheForFileUpdate invalidates cache for update on the given file
-func (fs *FileSystem) invalidateCacheForFileUpdate(path string) {
-	fs.cache.RemoveNegativeEntryCache(path)
-	fs.cache.RemoveEntryCache(path)
-
-	// modification doesn't affect to parent dir's modified time
-}
-
-// invalidateCacheForRemoveInternal invalidates cache for removal of the given file/dir
-func (fs *FileSystem) invalidateCacheForRemoveInternal(path string, recurse bool) {
-	var entry *Entry
-	if recurse {
-		entry = fs.cache.GetEntryCache(path)
-	}
-
-	fs.cache.RemoveEntryCache(path)
-	fs.cache.RemoveACLsCache(path)
-	fs.cache.RemoveMetadataCache(path)
-
-	if recurse && entry != nil {
-		if entry.Type == DirectoryEntry {
-			dirEntries := fs.cache.GetDirCache(path)
-			for _, dirEntry := range dirEntries {
-				// do it recursively
-				fs.invalidateCacheForRemoveInternal(dirEntry, recurse)
-			}
-		}
-	}
-
-	// remove dircache and dir acl cache even if it is a file or unknown, no harm.
-	fs.cache.RemoveDirCache(path)
-}
-
-// invalidateCacheForDirCreate invalidates cache for creation of the given dir
-func (fs *FileSystem) invalidateCacheForDirCreate(path string) {
-	fs.cache.RemoveNegativeEntryCache(path)
-
-	// parent dir's entry also changes
-	fs.cache.RemoveParentDirCache(path)
-	// parent dir's dir entry also changes
-	parentPath := util.GetIRODSPathDirname(path)
-	parentDirEntries := fs.cache.GetDirCache(parentPath)
-	if parentDirEntries != nil {
-		parentDirEntries = append(parentDirEntries, path)
-		fs.cache.AddDirCache(parentPath, parentDirEntries)
-	}
-}
-
-// invalidateCacheForDirRemove invalidates cache for removal of the given dir
-func (fs *FileSystem) invalidateCacheForDirRemove(path string, recurse bool) {
-	var entry *Entry
-	if recurse {
-		entry = fs.cache.GetEntryCache(path)
-	}
-
-	// we need to expunge all negatie entry caches under irodsDestPath
-	// since all sub-directories/files are also moved
-	fs.cache.RemoveAllNegativeEntryCacheForPath(path)
-
-	fs.cache.AddNegativeEntryCache(path)
-	fs.cache.RemoveEntryCache(path)
-	fs.cache.RemoveMetadataCache(path)
-
-	if recurse && entry != nil {
-		if entry.Type == DirectoryEntry {
-			dirEntries := fs.cache.GetDirCache(path)
-			for _, dirEntry := range dirEntries {
-				// do it recursively
-				fs.invalidateCacheForRemoveInternal(dirEntry, recurse)
-			}
-		}
-	}
-
-	fs.cache.RemoveDirCache(path)
-	fs.cache.RemoveACLsCache(path)
-
-	// parent dir's entry also changes
-	fs.cache.RemoveParentDirCache(path)
-	// parent dir's dir entry also changes
-	parentPath := util.GetIRODSPathDirname(path)
-	parentDirEntries := fs.cache.GetDirCache(parentPath)
-	if parentDirEntries != nil {
-		newParentDirEntries := []string{}
-		for _, dirEntry := range parentDirEntries {
-			if dirEntry != path {
-				newParentDirEntries = append(newParentDirEntries, dirEntry)
-			}
-		}
-		fs.cache.AddDirCache(parentPath, newParentDirEntries)
-	}
-}
-
-// invalidateCacheForFileCreate invalidates cache for creation of the given file
-func (fs *FileSystem) invalidateCacheForFileCreate(path string) {
-	fs.cache.RemoveNegativeEntryCache(path)
-
-	// parent dir's entry also changes
-	fs.cache.RemoveParentDirCache(path)
-	// parent dir's dir entry also changes
-	parentPath := util.GetIRODSPathDirname(path)
-	parentDirEntries := fs.cache.GetDirCache(parentPath)
-	if parentDirEntries != nil {
-		parentDirEntries = append(parentDirEntries, path)
-		fs.cache.AddDirCache(parentPath, parentDirEntries)
-	}
-}
-
-// invalidateCacheForFileRemove invalidates cache for removal of the given file
-func (fs *FileSystem) invalidateCacheForFileRemove(path string) {
-	fs.cache.AddNegativeEntryCache(path)
-	fs.cache.RemoveEntryCache(path)
-	fs.cache.RemoveACLsCache(path)
-	fs.cache.RemoveMetadataCache(path)
-
-	// parent dir's entry also changes
-	fs.cache.RemoveParentDirCache(path)
-	// parent dir's dir entry also changes
-	parentPath := util.GetIRODSPathDirname(path)
-	parentDirEntries := fs.cache.GetDirCache(parentPath)
-	if parentDirEntries != nil {
-		newParentDirEntries := []string{}
-		for _, dirEntry := range parentDirEntries {
-			if dirEntry != path {
-				newParentDirEntries = append(newParentDirEntries, dirEntry)
-			}
-		}
-		fs.cache.AddDirCache(parentPath, newParentDirEntries)
-	}
-}
-
-// AddUserMetadata adds a user metadata
-func (fs *FileSystem) AddUserMetadata(user string, avuid int64, attName, attValue, attUnits string) error {
-	metadata := &types.IRODSMeta{
-		AVUID: avuid,
-		Name:  attName,
-		Value: attValue,
-		Units: attUnits,
-	}
-
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	err = irods_fs.AddUserMeta(conn, user, metadata)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// DeleteUserMetadata deletes a user metadata
-func (fs *FileSystem) DeleteUserMetadata(user string, avuid int64, attName, attValue, attUnits string) error {
-	metadata := &types.IRODSMeta{
-		AVUID: avuid,
-		Name:  attName,
-		Value: attValue,
-		Units: attUnits,
-	}
-
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	err = irods_fs.DeleteUserMeta(conn, user, metadata)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ListUserMetadata lists all user metadata
-func (fs *FileSystem) ListUserMetadata(user string) ([]*types.IRODSMeta, error) {
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	metadataobjects, err := irods_fs.ListUserMeta(conn, user)
-	if err != nil {
-		return nil, err
-	}
-
-	return metadataobjects, nil
-}
-
-// GetTicketForAnonymousAccess gets ticket information for anonymous access
-func (fs *FileSystem) GetTicketForAnonymousAccess(ticket string) (*types.IRODSTicketForAnonymousAccess, error) {
-	conn, err := fs.session.AcquireConnection()
-	if err != nil {
-		return nil, err
-	}
-	defer fs.session.ReturnConnection(conn)
-
-	ticketInfo, err := irods_fs.GetTicketForAnonymousAccess(conn, ticket)
-	if err != nil {
-		return nil, err
-	}
-
-	return ticketInfo, err
 }
