+++ test/testcases/fs_api_test.go
@@ -9,7 +9,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cyverse/go-irodsclient/irods/common"
 	"github.com/cyverse/go-irodsclient/irods/connection"
 	"github.com/cyverse/go-irodsclient/irods/fs"
 	"github.com/cyverse/go-irodsclient/irods/session"
@@ -165,7 +164,7 @@ func testListIRODSDataObjects(t *testing.T) {
 		dataobjectPaths = append(dataobjectPaths, dataobject.Path)
 		assert.NotEmpty(t, dataobject.ID)
 
-		assert.Equal(t, string(common.GENERIC_DT), dataobject.DataType)
+		assert.Equal(t, string(types.GENERIC_DT), dataobject.DataType)
 
 		for _, replica := range dataobject.Replicas {
 			assert.NotEmpty(t, replica.Path)
@@ -201,7 +200,7 @@ func testListIRODSDataObjectsMasterReplica(t *testing.T) {
 		dataobjectPaths = append(dataobjectPaths, dataobject.Path)
 		assert.NotEmpty(t, dataobject.ID)
 
-		assert.Equal(t, string(common.GENERIC_DT), dataobject.DataType)
+		assert.Equal(t, string(types.GENERIC_DT), dataobject.DataType)
 
 		assert.Equal(t, 1, len(dataobject.Replicas))
 		for _, replica := range dataobject.Replicas {
@@ -242,7 +241,7 @@ func testGetIRODSDataObject(t *testing.T) {
 
 		assert.NotEmpty(t, dataobject.ID)
 
-		assert.Equal(t, string(common.GENERIC_DT), dataobject.DataType)
+		assert.Equal(t, string(types.GENERIC_DT), dataobject.DataType)
 
 		for _, replica := range dataobject.Replicas {
 			assert.NotEmpty(t, replica.Path)
@@ -278,7 +277,7 @@ func testGetIRODSDataObjectMasterReplica(t *testing.T) {
 
 		assert.NotEmpty(t, dataobject.ID)
 
-		assert.Equal(t, string(common.GENERIC_DT), dataobject.DataType)
+		assert.Equal(t, string(types.GENERIC_DT), dataobject.DataType)
 
 		assert.Equal(t, 1, len(dataobject.Replicas))
 		for _, replica := range dataobject.Replicas {
